move sequelize init to separate file remove dependency loop
