+++ src/models/index.ts
@@ -1,23 +1,4 @@
-/* eslint-disable no-magic-numbers, @typescript-eslint/no-explicit-any, import/no-cycle, import/first, no-console */
-
-import { Options, Sequelize } from 'sequelize';
-import { db, prod } from '../config';
-
-(Sequelize as any).postgres.DECIMAL.parse = parseFloat;
-
-export const sequelize = new Sequelize(db.database, db.user, db.password, {
-  logging: prod ? false : console.log,
-  host: db.host,
-  dialect: 'postgres',
-  pool: {
-    max: 5,
-    min: 0,
-    acquire: 30000,
-    idle: 10000,
-  },
-} as Options);
-
-// to avoid dependency loop
+import { sequelize } from '../lib/sequelize';
 import { Blocked } from './blocked';
 import { Message } from './message';
 
@@ -28,4 +9,4 @@ const initAssociations = () => {
 
 initAssociations();
 
-export { Blocked, Message };
+export { Blocked, Message, sequelize };
